Refactored freehand's keyDownCallback and keyUpCallback to utils/freehand/keysHeld.js
