+++ src/imageTools/freehand.js
@@ -13,6 +13,7 @@ import { addToolState, getToolState } from '../stateManagement/toolState.js';
 import { setToolOptions, getToolOptions } from '../toolOptions.js';
 
 // Freehand tool libraries
+import { keyDownCallback, keyUpCallback } from '../util/freehand/keysHeld.js';
 import dragObject from '../util/freehand/dragObject.js';
 import dropObject from '../util/freehand/dropObject.js';
 import insertOrDelete from '../util/freehand/insertOrDelete.js';
@@ -499,6 +500,7 @@ function checkHandlesPolygonMode (data) {
   const mousePoint = config.mouseLocation.handles.start;
   const dataHandles = data.handles;
   let invalidHandlePlacement = false;
+
   data.canComplete = false;
 
   const mouseAtOriginHandle = (external.cornerstoneMath.point.distance(dataHandles[0], mousePoint) < config.spacing);
@@ -726,74 +728,6 @@ function mouseHover (eventData, toolData) {
   }
 
   return imageNeedsUpdate;
-}
-
-const keyCodes = {
-    SHIFT: 16,
-    CTRL: 17,
-    ALT: 18,
-};
-
-/**
-* Event handler for KEY_DOWN event.
-*
-* @param {Object} e - The event.
-*/
-function keyDownCallback (e) {
-  const eventData = e.detail;
-  const config = freehand.getConfiguration();
-  const keyCode = eventData.keyCode;
-  let imageNeedsUpdate = false;
-
-  if (keyCode === keyCodes.CTRL) {
-    config.keyDown.ctrl = true;
-    imageNeedsUpdate = true;
-  }
-
-  if(keyCode === keyCodes.SHIFT) {
-    config.keyDown.shift = true;
-  }
-
-  if(keyCode === keyCodes.ALT) {
-    config.keyDown.alt = true;
-  }
-
-  if (imageNeedsUpdate) {
-    // Force onImageRendered to fire
-    external.cornerstone.updateImage(eventData.element);
-  }
-}
-
-
-
-/**
-* Event handler for KEY_UP event.
-*
-* @param {Object} e - The event.
-*/
-function keyUpCallback (e) {
-  const eventData = e.detail;
-  const config = freehand.getConfiguration();
-  const keyCode = eventData.keyCode;
-  let imageNeedsUpdate = false;
-
-  if (keyCode === keyCodes.CTRL) {
-    config.keyDown.ctrl = false;
-    imageNeedsUpdate = true;
-  }
-
-  if(keyCode === keyCodes.SHIFT) {
-    config.keyDown.shift = false;
-  }
-
-  if(keyCode === keyCodes.ALT) {
-    config.keyDown.alt = false;
-  }
-
-  if (imageNeedsUpdate) {
-    // Force onImageRendered to fire
-    external.cornerstone.updateImage(eventData.element);
-  }
 }
 
 /**
