+++ src/inputSources/keyboardInput.js
@@ -10,8 +10,6 @@ function keyPress (e) {
   const element = e.currentTarget;
   const enabledElement = cornerstone.getEnabledElement(element);
 
-  console.log('keypressed');
-
   if (!enabledElement.image) {
     return;
   }
